Added PropertyFileReader
 PropertyFileReader to read the website URL from an external configuration file instead of writing it directly in the code.
Created an object of PropertyFileReader  then use it to fetch the value for the key "url" from a file named "config.properties". 
This value is stored in the variable baseURL and can be used later to open the website. By doing this  if the URL ever changes  don’t have to update  Java code just update the value in the properties file.